feat(#33): cover create group and restructure files
