+++ internal/web/v1/groups/groupcontroller.go
@@ -10,6 +10,13 @@ import (
 	"github.com/eurofurence/reg-room-service/internal/web/common"
 )
 
+// Controller implements methods which satisfy the endpoint format
+// in the `common` package.
+type Controller struct {
+	ctrl groupservice.Service
+}
+
+// InitRoutes creates the Controller instance and sets up all routes on it.
 func InitRoutes(router chi.Router, ctrl groupservice.Service) {
 	h := &Controller{
 		ctrl: ctrl,
